+++ NetIdMobileSdk/NetIdMobileSdk/api/NetIdConfig.swift
@@ -41,8 +41,6 @@ public struct PermissionLayerConfig {
 }
 
 public struct NetIdConfig {
-    /// Name of the host that acts as an SSO broker, if not set defaults to broker.netid.de.
-    public var host: String
     /// The client id of this application. You need to retrieve it from the netID developer portal.
     public var clientId: String
     /// Redirect URI for your application.  You need to retrieve it from the netID developer portal.
@@ -56,7 +54,6 @@ public struct NetIdConfig {
 
     /// Initialize the SDK. This is the first thing to do.
     public init(host: String, clientId: String, redirectUri: String, claims: [String: String]?, loginLayerConfig: LoginLayerConfig?, permissionLayerConfig: PermissionLayerConfig?) {
-        self.host = host
         self.clientId = clientId
         self.redirectUri = redirectUri
         self.claims = claims